Cool sound remains, but less obnoxious
Sound category changed to BLOCKS, volume reduced by default.

This will help make this block much less obnoxious on servers. As of right now, it is very loud and annoying for others on servers, especially when disenchanting multiple items in a row.

Unchanged but suspected issue:
ITEM_TOTEM_USE is playing for every client that has the chunks loaded as well, perhaps this is due to client.player.playSound calling on every client and not just the block user?